Add login, logout utils for e2e tests and use them
