+++ e2e/integration/auth.spec.js
@@ -1,7 +1,8 @@
-import { API_WAIT_MS, DB_PROFILES, loadDatabaseProfile } from "../utils";
+import { DB_PROFILES, loadDatabaseProfile, login, logout } from "../utils";
 
 describe("authentication", () => {
   beforeEach(() => {
+    // clean auth state
     cy.clearCookies();
     cy.clearLocalStorage();
   });
@@ -21,15 +22,9 @@ describe("authentication", () => {
       // credentials
       const user = db.user[0];
       const { email, first_name, last_name } = user;
-      const password = email.split("@")[0];
 
-      // fill in credentials
-      cy.get("input[type=email]").type(email);
-      cy.get("input[type=password]").type(password);
-      cy.get("button[type=submit]").click();
-
-      // await logged in page
-      cy.wait(API_WAIT_MS);
+      // fill in credentials form and submit
+      login(email);
 
       // sample table present?
       cy.get(`table#sampleTable`).should("exist");
@@ -45,22 +40,13 @@ describe("authentication", () => {
 
       // credentials
       const user = db.user[0];
-      const { email, first_name, last_name } = user;
-      const password = email.split("@")[0];
+      const { email } = user;
 
-      // fill in credentials
-      cy.get("input[type=email]").type(email);
-      cy.get("input[type=password]").type(password);
-      cy.get("button[type=submit]").click();
+      // fill in credentials form and submit
+      login(email);
 
-      // await logged in page
-      cy.wait(API_WAIT_MS);
-
-      // logout
-      cy.contains("Logout").click();
-
-      // await login page again
-      cy.wait(API_WAIT_MS);
+      // click logout button
+      logout();
 
       // credentials form?
       cy.get("input[type=email]").should("exist");
